pyproject.toml
[tool.poetry]
name = "pro"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
twisted = "^20.3.0"
youtube-dl = "^2020.12.2"
configparser = "^5.0.1"
httplib2 = "^0.18.1"
pure-sasl = "^0.6.2"
pystick = "^0.1.6"
tornado = "^6.1"
gtts = "^2.2.1"
humanfriendly = "^9.0"
python-dotenv = "^0.15.0"
var-dump = "^1.2"
python-axolotl-curve25519 = "^0.4.1"
pafy = "^0.5.5"
pytz = "^2020.4"
humanize = "^3.1.0"
qrcode = "^6.1"
mechanize = "^0.4.5"
lxml = "^4.6.2"
Pillow = "^8.0.1"
rsa = "^4.6"
bs4 = "^0.0.1"
timeago = "^1.0.14"
astroid = "^2.4.2"
pytest-localserver = "^0.5.0"
google-auth = "^1.23.0"
tqdm = "^4.54.0"
progressbar = "^2.5"
httpx = "0.13.3"
wikipedia = "^1.4.0"
googletrans = "^3.0.0"
aiohttp = "^3.7.3"
uvloop = "^0.14.0"
pycrypto = "^2.6.1"
goslate = "^1.5.1"
dropbox = "^11.0.0"
pycryptodome = "^3.10.1"
xxhash = "^2.0.0"
thrift = "^0.13.0"
IMDbPY = "^2021.4.18"
toro = "^1.0.1"
ply = "^3.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
